attach finish commit

+++ src/main/java/org/example/controller/AuthorizationController.java
@@ -3,15 +3,12 @@ package org.example.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
-import lombok.Value;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.tomcat.util.http.parser.Authorization;
 import org.example.dto.AuthorizationResponseDTO;
 import org.example.dto.LoginDTO;
 import org.example.dto.RegistrationDTO;
 import org.example.service.AuthorizationService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,7 +28,7 @@ public class AuthorizationController {
     @Operation(summary = "Registration", description = "Api for profile registration")
     public ResponseEntity<String> registration(@Valid @RequestBody RegistrationDTO registrationDTO){
         String registration = authorizationService.registration(registrationDTO);
-        log .info("Registration name = {} email = {}",registrationDTO.getName(), registrationDTO.getEmail());
+        log.info("Registration name = {} email = {}",registrationDTO.getName(), registrationDTO.getEmail());
         return ResponseEntity.ok().body(registration);
     }
 
@@ -42,7 +39,7 @@ public class AuthorizationController {
     }
 
     @GetMapping("/verification/{userId}")
-    public ResponseEntity<String> verification(@PathVariable("userId") Integer userId) {
+    public ResponseEntity<String> verification(@PathVariable("userId") Long userId) {
         String body = authorizationService.authorizationVerification(userId);
         return ResponseEntity.ok().body(body);
     }
